feat: Add SlowCommand class, move CommandClasses to a commands folder.
