+++ src/CommandHandler.ts
@@ -3,8 +3,8 @@ import {EventEmitter} from 'events';
 import {promises as fsPromises} from 'fs';
 import {join} from 'path';
 import {AdvancedClient} from './classes/AdvancedClient';
-import {Command} from './classes/Command';
-import {CommandHandlerError} from './classes/CommandHandlerError';
+import {Command} from './classes/commands/Command.js';
+import {CommandHandlerError} from './classes/commands/CommandHandlerError.js';
 import {Event} from './classes/Event';
 import * as defaultCommands from './defaults/commands/index';
 import * as defaultEvents from './defaults/events/index';
