+++ src/defaults/events/message.ts
@@ -1,6 +1,6 @@
 import {Message} from 'discord.js';
-import {Tag} from '../../classes/Command';
-import {CommandContext} from '../../classes/CommandContext.js';
+import {Tag} from '../../classes/commands/Command.js';
+import {CommandContext} from '../../classes/commands/CommandContext.js';
 import {Event} from '../../classes/Event';
 import {EventContext} from '../../classes/EventContext.js';
 import {CommandHandler} from '../../CommandHandler';
