+++ src/index.ts
@@ -3,9 +3,9 @@ export {dayjs};
 
 export * from 'discord.js-better-embed';
 export * from './CommandHandler';
-export * from './classes/Command';
+export * from './classes/commands/Command.js';
 export * from './classes/AdvancedClient';
-export * from './classes/CommandHandlerError';
+export * from './classes/commands/CommandHandlerError.js';
 export * from './classes/Event';
 export * from './utils/argError';
 export * from './utils/codeError';
